# openapi.json
{
  "components": {
...
    "securitySchemes": {
      "SupabaseJWTBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\n## AI Fitness Companion Backend API\n\nThe Slow Burn API provides a comprehensive backend for an AI-powered fitness companion application.\nUsers develop a deepening relationship with an AI companion as they log workouts and achieve their fitness goals.\n\n### Key Features:\n- 🔐 **JWT Authentication** via Supabase\n- 💪 **Workout Plan Management** with versioning support\n- 🏋️ **Exercise Library** (coming in Sprint 5)\n- 🤖 **AI Chat Integration** with Google Gemini (coming in Sprint 6)\n- 📈 **Progress Tracking** and affinity scoring\n\n### Current Sprint: 4 (Workout Logging)\nFor more details, see the [project documentation](https://github.com/yourusername/slow-burn).\n    ",
    "title": "Slow Burn API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workouts/{workout_id}": {
      "delete": {
        "description": "Delete a workout session and all associated sets.\n\nPerforms a hard delete of the workout session and all its associated sets.\nOnly the workout owner can delete their workouts. This operation is\npermanent and cannot be undone.\n\nArgs:\n    workout_id: The UUID of the workout to delete\n    jwt_payload: The authenticated user's JWT payload\n    repo: The workouts repository instance\n\nReturns:\n    None (204 No Content on success)\n\nRaises:\n    HTTPException:\n        - 403 if user doesn't own the workout\n        - 404 if workout doesn't exist\n        - 500 if deletion fails due to database errors",
        "operationId": "delete_workout_api_v1_workouts__workout_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workout Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Workout deleted successfully"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticated"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "You do not have permission to delete this workout"
                }
              }
            },
            "description": "User does not own this workout"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Workout not found"
                }
              }
            },
            "description": "Workout not found"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "path",
                        "workout_id"
                      ],
                      "msg": "value is not a valid uuid",
                      "type": "type_error.uuid"
                    }
                  ]
                }
              }
            },
            "description": "Invalid workout ID format"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Failed to delete workout"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Workout deletion endpoint not yet implemented"
                }
              }
            },
            "description": "Not implemented"
          }
        },
        "security": [
          {
            "SupabaseJWTBearer": []
          }
        ],
        "summary": "Delete a workout session",
        "tags": [
          "workouts"
        ]
      },
      "get": {
        "description": "Retrieve a specific workout session with all its sets.\n\nReturns the detailed workout data for the specified workout_id, including\nall sets performed, exercise information, and wellness metrics. Access is\nrestricted to the workout owner via RLS policies.\n\nThe response includes:\n- Full workout session details\n- All sets with exercise names and performance metrics\n- Calculated statistics (total volume, duration)\n- Wellness tracking data\n\nArgs:\n    workout_id: The UUID of the workout to retrieve\n    jwt_payload: The authenticated user's JWT payload\n    repo: The workouts repository instance\n\nReturns:\n    WorkoutDetailResponseModel containing the workout details with all sets\n\nRaises:\n    HTTPException:\n        - 404 if workout doesn't exist or user lacks permission\n        - 500 if retrieval fails due to database errors",
        "operationId": "get_workout_by_id_api_v1_workouts__workout_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workout Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponseModel"
                }
              }
            },
            "description": "Workout session retrieved successfully"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticated"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Workout not found"
                }
              }
            },
            "description": "Workout not found or access denied"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "path",
                        "workout_id"
                      ],
                      "msg": "value is not a valid uuid",
                      "type": "type_error.uuid"
                    }
                  ]
                }
              }
            },
            "description": "Invalid workout ID format"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Failed to retrieve workout"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Workout retrieval endpoint not yet implemented"
                }
              }
            },
            "description": "Not implemented"
          }
        },
        "security": [
          {
            "SupabaseJWTBearer": []
          }
        ],
        "summary": "Get a specific workout session by ID",
        "tags": [
          "workouts"
        ]
      }
    },
...
  },
  "security": [
    {
      "SupabaseJWTBearer": []
    }
  ],
  "servers": [
    {
      "description": "Default API server",
      "url": "http://localhost:8000"
    }
  ]
}